ai_server_pdf/requirements.txt
fastapi==0.105.0
python-dotenv==1.0.0
google-generativeai==0.3.2
PyPDF2==3.0.1
pydantic==2.5.0
uvicorn==0.24.0
requests==2.31.0
python-multipart==0.0.6
